# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A modern Go-based authentication system with magic link functionality using Clean Architecture principles.\nThis API provides secure authentication mechanisms including magic link authentication, JWT token management, and user profile management.",
        "title": "Trading Alchemist API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Trading Alchemist Team",
            "email": "team@tradingalchemist.dev"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/magic-link": {
            "post": {
                "description": "Sends a magic link to the specified email address for passwordless authentication. The magic link will be valid for the configured TTL period (default 15 minutes).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Send magic link",
                "parameters": [
                    {
                        "description": "Email address to send magic link to",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_application_auth.SendMagicLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Magic link sent successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/trading-alchemist_internal_application_auth.SendMagicLinkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid email address or validation error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too many requests - rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/verify": {
            "post": {
                "description": "Verifies a magic link token from an email link and returns a JWT access token if valid. The magic link token is consumed and cannot be used again.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify magic link",
                "parameters": [
                    {
                        "description": "Magic link token to verify",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_application_auth.VerifyMagicLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Magic link verified successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/trading-alchemist_internal_application_auth.VerifyMagicLinkResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid token format or token missing",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid, expired, or already used token",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Token not found",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "trading-alchemist_internal_application_auth.SendMagicLinkRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "description": "Email address to send the magic link to",
                    "type": "string"
                },
                "purpose": {
                    "description": "Purpose of the magic link (login, email_verification, password_reset)",
                    "type": "string"
                }
            }
        },
        "trading-alchemist_internal_application_auth.SendMagicLinkResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Response message",
                    "type": "string"
                },
                "sent": {
                    "description": "Whether the email was sent successfully",
                    "type": "boolean"
                }
            }
        },
...
        "trading-alchemist_internal_application_auth.UserResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "description": "Avatar URL",
                    "type": "string"
                },
                "created_at": {
                    "description": "Account creation timestamp",
                    "type": "string"
                },
                "display_name": {
                    "description": "Display name",
                    "type": "string"
                },
                "email": {
                    "description": "Email address",
                    "type": "string"
                },
                "email_verified": {
                    "description": "Whether email is verified",
                    "type": "boolean"
                },
                "first_name": {
                    "description": "First name",
                    "type": "string"
                },
                "full_name": {
                    "description": "Full name (computed from first and last name)",
                    "type": "string"
                },
                "id": {
                    "description": "User ID",
                    "type": "string"
                },
                "is_active": {
                    "description": "Whether user is active",
                    "type": "boolean"
                },
                "last_name": {
                    "description": "Last name",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string"
                }
            }
        },
        "trading-alchemist_internal_application_auth.VerifyMagicLinkRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "Magic link token to verify",
                    "type": "string"
                }
            }
        },
        "trading-alchemist_internal_application_auth.VerifyMagicLinkResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "JWT access token",
                    "type": "string"
                },
                "expires_in": {
                    "description": "Token expiration in seconds",
                    "type": "integer"
                },
                "token_type": {
                    "description": "Token type",
                    "type": "string"
                },
                "user": {
                    "description": "User information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/trading-alchemist_internal_application_auth.UserResponse"
                        }
                    ]
                }
            }
        },
...
        "trading-alchemist_internal_presentation_responses.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Error code\nexample: VALIDATION_ERROR",
                    "type": "string"
                },
                "details": {
                    "description": "Optional error details\nexample: The email field must be a valid email address",
                    "type": "string"
                },
                "message": {
                    "description": "Error message\nexample: Invalid email address",
                    "type": "string"
                }
            }
        },
        "trading-alchemist_internal_presentation_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorDetail"
                        }
                    ]
                },
                "success": {
                    "description": "Success indicator (always false for errors)\nexample: false",
                    "type": "boolean"
                }
            }
        },
        "trading-alchemist_internal_presentation_responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data"
                },
                "message": {
                    "description": "Optional success message\nexample: Operation completed successfully",
                    "type": "string"
                },
                "success": {
                    "description": "Success indicator\nexample: true",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token. Usage: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}